viewer-request/requirements.txt
PyJWT==1.7.1
cryptography==2.8
